Factored probability and three_d constructs out of topics/
